refactor: [#58] 북마크 토글 컴포넌트 분리
