+++ src/components/shared/Photo.tsx
@@ -1,8 +1,6 @@
 import styled from 'styled-components';
 import { PhotoItem } from '../../types/photos';
-import useToggleBookmark from '../../hooks/useToggleBookmark';
-import StyledHeartLineIcon from '../../assets/icons/StyledHeartLineIcon';
-import StyledHeartFillIcon from '../../assets/icons/StyledHeartFillIcon';
+import BookmarkToggle from '../BookmarkToggle';
 import { device } from '../../styles/theme';
 
 interface PhotoProps {
@@ -14,25 +12,10 @@ interface PhotoProps {
 }
 
 function Photo({ photo, url, alt, onClick, isSkeleton = false }: PhotoProps) {
-  const { handleToggleBookmark, getBookmarkStatus } = useToggleBookmark();
-
-  const handleBookmarkClick = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    handleToggleBookmark(photo);
-  };
-
   return (
     <Container onClick={onClick} $isSkeleton={isSkeleton}>
       <Thumbnail src={url} alt={alt} />
-      {!isSkeleton && (
-        <IconContainer onClick={(e) => handleBookmarkClick(e)}>
-          {getBookmarkStatus(photo.id) ? (
-            <StyledHeartFillIcon colorname="red" />
-          ) : (
-            <StyledHeartLineIcon colorname="white" />
-          )}
-        </IconContainer>
-      )}
+      {!isSkeleton && <BookmarkToggle photo={photo} />}
     </Container>
   );
 }
@@ -65,19 +48,4 @@ const Thumbnail = styled.img`
   width: 100%;
   height: 100%;
   object-fit: cover;
-`;
-
-const IconContainer = styled.div`
-  position: absolute;
-  width: 20px;
-  height: 20px;
-  right: 4px;
-  bottom: 4px;
-
-  @media ${device.tablet} {
-    width: 32px;
-    height: 32px;
-    right: 8px;
-    bottom: 8px;
-  }
 `;
